pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sexism-distillation"
version = "0.1.0"
description = "Adaptive student distillation toolkit for sexism classification"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.1",
  "transformers>=4.40",
  "pyyaml>=6.0",
]

[tool.setuptools]
packages = ["sexism_distillation"]

[tool.setuptools.package-dir]
"sexism_distillation" = "sexism_distillation"
